requirements.txt
alive_progress==3.1.5
mutagen==1.47.0
Pillow==11.0.0
python-dotenv==1.0.1
Requests==2.32.3
ytmusicapi==1.8.2
